initial commit of old sources
this is an ancient (circa-2005), small and fast expat-based xml file
well-formedness checker. it examines (and, optionally, moves) xml
files and via $? returns whether or not the file is well-formed.

this arose from having a system where a *gigantic* number of xml
files were added to a deep filesystem hierarchy by a number of
concurrent data import processes.  sometimes the files would get
partially written or otherwise corrupted, causing down-stream errors.
we needed a way to quickly find and remove them, so this in concert
with find(1) and such got the job done.